SObject template implementation
